Show db only networks when choosing network to display
