+++ app/src/main/java/com/wifiheatmap/wifiheatmap/MainDrawerAdapter.kt
@@ -1,52 +1,17 @@
 package com.wifiheatmap.wifiheatmap
 
-import android.net.wifi.ScanResult
-import android.net.wifi.WifiManager
-import android.os.Bundle
 import android.view.*
 import android.widget.ImageView
 import android.widget.TextView
-import android.widget.Toast
-import androidx.core.os.bundleOf
-import androidx.core.view.isVisible
 import androidx.drawerlayout.widget.DrawerLayout
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.lifecycle.ViewModelProviders
-import androidx.recyclerview.widget.RecyclerView
-import com.wifiheatmap.wifiheatmap.room.Network
-import kotlinx.android.synthetic.main.network_selection_item.view.*
 import kotlinx.android.synthetic.main.network_selection_item_drawer.view.*
 
-class MainDrawerAdapter : RecyclerView.Adapter<MainDrawerAdapter.NetworkHolder>() {
+class MainDrawerAdapter : NetworkListAdapter() {
 
-    public lateinit var drawerLayout: DrawerLayout
+    lateinit var drawerLayout: DrawerLayout
 
     // this is being set from the MainActivity when this MainDrawerAdapter is created.
-    public lateinit var mapsViewModel: MapsViewModel
-
-    class NetworkHolder(val networkView: View) : RecyclerView.ViewHolder(networkView)
-
-    private data class NetworkData(val networkDB: Network? = null, val scanResult: ScanResult?)
-
-    private var networks: MutableList<NetworkData>? = null
-
-    fun setNetworks(scanResults: List<ScanResult>) {
-        if (this.networks == null) {
-            this.networks = mutableListOf(NetworkData(null, null))
-        }
-        this.networks!!.clear()
-        for (wifi: ScanResult in scanResults) {
-            var matchedNetwork: Network? = null
-            if (matchedNetwork != null) {
-                this.networks!!.add(NetworkData(matchedNetwork, wifi))
-            } else {
-                this.networks!!.add(NetworkData(null, wifi))
-            }
-        }
-
-        notifyDataSetChanged()
-    }
+    lateinit var mapsViewModel: MapsViewModel
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): NetworkHolder {
         val view = (LayoutInflater.from(parent.context).inflate(
@@ -56,11 +21,6 @@ class MainDrawerAdapter : RecyclerView.Adapter<MainDrawerAdapter.NetworkHolder>(
         ))
         return NetworkHolder(view)
     }
-
-    override fun getItemCount(): Int {
-        return networks?.size ?: 0
-    }
-
 
     override fun onBindViewHolder(holder: NetworkHolder, position: Int) {
 
@@ -79,14 +39,7 @@ class MainDrawerAdapter : RecyclerView.Adapter<MainDrawerAdapter.NetworkHolder>(
             // Display the name of the network from whichever source is non-null
             val ssidLabel = holder.networkView.findViewById<TextView>(R.id.ssid_label_drawer)
 
-            if (network.scanResult != null) {
-                ssidLabel.text = network.scanResult.SSID
-            }
-
-            // might want to check the current networks in this recycler view and
-            // strikethrough all the ones that are blacklisted within the database.
-
-
+            ssidLabel.text = network.scanResult?.SSID ?: network.networkDB!!.ssid
 
             // display the strength of the current network as an icon.
             val strength = holder.networkView.findViewById<ImageView>(R.id.wifi_strength_drawer)
@@ -94,22 +47,6 @@ class MainDrawerAdapter : RecyclerView.Adapter<MainDrawerAdapter.NetworkHolder>(
             strength.setBackgroundResource(getWifiStrengthIcon(network.scanResult?.level))
 
         }
-    }
-
-    private fun getWifiStrengthIcon(strength: Int?) : Int {
-        return if (strength != null) {
-            when (WifiManager.calculateSignalLevel(strength, 5)) {
-                0 -> R.drawable.ic_signal_wifi_0_bar_black_24dp
-                1 -> R.drawable.ic_signal_wifi_1_bar_black_24dp
-                2 -> R.drawable.ic_signal_wifi_2_bar_black_24dp
-                3 -> R.drawable.ic_signal_wifi_3_bar_black_24dp
-                4 -> R.drawable.ic_signal_wifi_4_bar_black_24dp
-                else -> R.drawable.ic_signal_wifi_off_black_24dp
-            }
-        } else {
-            R.drawable.ic_signal_wifi_off_black_24dp // default value
-        }
-        return 0
     }
 
 }
